Moved router_service and handlers to different folders.
Handlers are now in agent_endpoints.
The Master route is in its own /routes folder. Getting master route seemed unnecessary, so it was placed with the endpoint. This way, the endpoint basically does:
1. Input validation
2. Session retrieval
3. Build the router using route functions from /routes
4. Run the response.

If you feel it is necessary to abstract router construction in its own folder, feel free to add it back in router_service, although i think it makes it more complicated than necessary, since the router and routes itself are kind of the "service".